alpha v0

+++ database/migrations/2023_08_22_201259_create_areas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nombre');
             $table->string('tproducto');
             $table->string('direccion');
-            $table->integer('telefono');
+            $table->double('telefono');
             $table->string('correo');
             $table->timestamps();
         });
